+++ src/services/clock.rs
@@ -10,21 +10,22 @@ use std::sync::{Arc, Mutex};
 use std::thread;
 use std::time::Duration;
 
-#[derive(Debug)]
 pub struct Clock {
+    service_id: String,
     interval: Duration,
     counter: u64,
 }
 
 #[derive(Deserialize, Debug, Clone)]
-pub struct ClockSettings {
+pub struct Settings {
     /// Interval in milliseconds.
     pub interval_ms: Option<u64>,
 }
 
 impl Clock {
-    pub fn new(settings: &ClockSettings) -> Clock {
+    pub fn new(service_id: &str, settings: &Settings) -> Clock {
         Clock {
+            service_id: service_id.into(),
             interval: Duration::from_millis(settings.interval_ms.unwrap_or(1000)),
             counter: 0,
         }
@@ -32,18 +33,11 @@ impl Clock {
 }
 
 impl crate::services::Service for Clock {
-    fn spawn(
-        mut self: Box<Self>,
-        service_id: String,
-        _db: Arc<Mutex<Db>>,
-        tx: Sender<Reading>,
-        _rx: Receiver<Reading>,
-    ) -> Vec<JoinHandle> {
-        let sensor = service_id.clone();
-        vec![threading::spawn(service_id, move || loop {
+    fn spawn(mut self: Box<Self>, _db: Arc<Mutex<Db>>, tx: Sender<Reading>, _rx: Receiver<Reading>) -> Vec<JoinHandle> {
+        vec![threading::spawn(self.service_id.clone(), move || loop {
             #[rustfmt::skip]
             tx.send(Reading {
-                sensor: sensor.clone(),
+                sensor: self.service_id.clone(),
                 value: Value::Counter(self.counter),
                 timestamp: Local::now(),
                 is_persisted: true,